% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{analyticlm}
\alias{analyticlm}
\title{Analytic solution for least squares}
\usage{
analyticlm(X, y)
}
\arguments{
\item{X}{attribute matrix with intercept coefficient}

\item{y}{target vector}
}
\value{
list with theta vector
}
\description{
Analytic solution for least squares
}
