% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{analyticridge}
\alias{analyticridge}
\title{Functions associated with ridge regression
Analytic solution for Ridge Regression}
\usage{
analyticridge(X, y, lambda)
}
\arguments{
\item{X}{attribute matrix with intercept coefficient}

\item{y}{target vector}

\item{lambda}{regularization parameter}
}
\value{
list with theta vector
}
\description{
Functions associated with ridge regression
Analytic solution for Ridge Regression
}
