% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes}
\alias{bayes}
\title{Bayesian object}
\usage{
bayes(lambda, kernel = "gaussian")
}
\arguments{
\item{lambda}{kernel bin width}

\item{kernel, }{string name of kernel to use}
}
\value{
s3 object, bayes
}
\description{
Used for Bayesian classification with kernel
density estimation
}
