% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{bernoulli_sim}
\alias{bernoulli_sim}
\title{Simulate Bernoulli distribution from HW3}
\usage{
bernoulli_sim()
}
\value{
list, attribute matrix, target vector
}
\description{
Simulate Bernoulli distribution from HW3
}
