% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradesc.R
\name{bgd}
\alias{bgd}
\title{Gradient descent algorithms}
\usage{
bgd(alpha, X, y, grad, N, precision = 1e-04, ...)
}
\arguments{
\item{alpha}{learning rate constant}

\item{X}{matrix of predictors including 1 for beta 0}

\item{y}{vector of dependent variable}

\item{grad}{gradient function}

\item{N}{maximum number of iterations}

\item{precision}{constant used for approximation condition}
}
\value{
list object with theta, convergence status, and number of iterations
}
\description{
Meant for use with loss functions in supml
Batch Gradient Descent
}
\examples{
x <- runif(1000, -5, 5)
y <- x + rnorm(1000) + 3
X <- cbind(1, matrix(x))
h <- bgd(0.1, X, y, least_squares_gradient, 1000)
h$theta
h$converged == TRUE
h$iterations

}
