% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom_mean}
\alias{binom_mean}
\title{Mean of binomial variable}
\usage{
binom_mean(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
mean value of success for n trials
}
\description{
Mean of binomial variable
}
\examples{
binom_mean(1,1)
binom_mean(10, 0.35)

}
