% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom_sd}
\alias{binom_sd}
\title{Standard deviation of binomial variable}
\usage{
binom_sd(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
standard deviation of success after n trials
}
\description{
Standard deviation of binomial variable
}
\examples{
binom_sd(1, 1)
binom_sd(10, 0.35)

}
