% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binomcdf}
\alias{binomcdf}
\title{Compute the binomial cumulative density function}
\usage{
binomcdf(n, k, p)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}

\item{p}{probability of success for an independent trial}
}
\value{
binomial cumulative probability of k successes
}
\description{
Compute the binomial cumulative density function
}
\examples{
binomcdf(1,1,1)
binomcdf(7,4,0.35)

}
