% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binompdf}
\alias{binompdf}
\title{Compute the binomial probability density function}
\usage{
binompdf(n, k, p)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}

\item{p}{probability of success for an independent trial}
}
\value{
The probability of having exactly k successes
}
\description{
Compute the binomial probability density function
}
\examples{
binompdf(1,1,1)
binompdf(7, 4, 0.35) # probability of making 7 out of 4 free throws in bball

}
