% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{categoricalProbs.bayes}
\alias{categoricalProbs.bayes}
\title{Conditional probability fitting step for Categorical data}
\usage{
\method{categoricalProbs}{bayes}(object, X, y)
}
\value{
"n|j|label"
}
\description{
Conditional probability fitting step for Categorical data
}
