% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{combs}
\alias{combs}
\title{Discrete random variables -- binomial probability
Compute the number of combinations of for n, k}
\usage{
combs(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
The number of combinations for n choose k
}
\description{
Discrete random variables -- binomial probability
Compute the number of combinations of for n, k
}
