% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{confusionMatrix.bayes}
\alias{confusionMatrix.bayes}
\title{Confusion matrix for reporting accuracy}
\usage{
\method{confusionMatrix}{bayes}(object, Y_true, Y_predict)
}
\arguments{
\item{object}{bayes s3 object}

\item{Y_true}{boolean matrix of correct values}

\item{Y_predict}{boolean matrix of predicted values}
}
\value{
2 by 2 confusion matrix
}
\description{
Using an accuracy score when classes are imbalanced
can be really deceiving. A confusion matrix can be
used as an alternative measure of performance. This
confusion matrix expects boolean spatial matrices
for Y. It uses a 'resolve' matrix to exclude ineligible
predictions.
}
