% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{continuousProbs.bayes}
\alias{continuousProbs.bayes}
\title{Conditional probability fitting step for Continuous data}
\usage{
\method{continuousProbs}{bayes}(object, X, y, continuous_cols)
}
\arguments{
\item{object}{bayesian s3 object with mappings set}

\item{X}{entire feature matrix}

\item{y}{target vector, factor}

\item{continuous_cols}{int vector of column positions}
}
\value{
mean for each column j given a y class label; 'mu|j|y'
}
\description{
Only 'gaussian' kernel currently supported. Gaussian kernel
requies mean mu given y in addition to the priors for y. The
Gaussian kernel standard deviation is specified by hyperparameter
lambda and not required for conditional probability computation.
Priors for y are computed for the model rather than duplicated for
each continuous feature j.
}
\details{
This method should only be used for customizing your own
Bayesian classifier.
}
