% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{coord_cut}
\alias{coord_cut}
\title{Utility functions for working with Bayesian classifier
Create buckets given a vector x}
\usage{
coord_cut(x, nbreaks = 100)
}
\arguments{
\item{x}{numeric vector}

\item{nbreaks}{number of buckets for factor}
}
\value{
factor vector
}
\description{
Bayesian classification for a spatial matrix
requires that we normalize coordinates.
}
