% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{create_splits}
\alias{create_splits}
\title{For use with in generating training and test sets
Create training, test, and hold out splits}
\usage{
create_splits(data, train, test)
}
\arguments{
\item{data}{data frame object}

\item{train}{integer, number of training samples}

\item{test}{integer, number of test samples}
}
\value{
list of training, test and holdout samples
}
\description{
For use with in generating training and test sets
Create training, test, and hold out splits
}
