% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{data_simulator}
\alias{data_simulator}
\title{Simulate data for hw}
\usage{
data_simulator()
}
\value{
list, attribute matrix, target vector
}
\description{
Cleaner implementation from the TA's answer key
}
