% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{fit.bayes}
\alias{fit.bayes}
\title{Fit training data to Bayesian Classifier}
\usage{
\method{fit}{bayes}(object, X, y)
}
\arguments{
\item{object}{bayes s3 object}

\item{X}{feature matrix}

\item{y}{target vector}
}
\value{
obj updated bayes s3 object
}
\description{
We want to assign models to each class,
accounting for continuous and categorical
variables. We also handle spatial data.
Log priors are used to avoide issues with
floating point.
}
