% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{gauss_kern}
\alias{gauss_kern}
\title{Kernels used for localized estimation
Gaussian Kernel}
\usage{
gauss_kern(x, mu, lambda)
}
\arguments{
\item{x}{feature vector}

\item{mu}{mean}

\item{lambda}{adjustable parameter, usually standard deviation}
}
\value{
transformed feature vector using gaussian distribution
}
\description{
Solution using HT Ch. 6 Kernel Smoothing Methods (6.23)
}
