% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geometpdf}
\alias{geometpdf}
\title{Geometric probability density function for first success in x trials}
\usage{
geometpdf(p, x)
}
\arguments{
\item{p}{probability of success for independent trial}

\item{x}{number of trials, x >= 1}
}
\value{
probability of success after x number of trials
}
\description{
Geometric probability density function for first success in x trials
}
\examples{
geometpdf(1,1)
geometpdf(1/13, 5)

}
