% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{gridSearch.bayes}
\alias{gridSearch.bayes}
\title{Implement grid search for the bayes object}
\usage{
\method{gridSearch}{bayes}(obj, X, y, grid, score)
}
\arguments{
\item{obj}{s3 object bayes}
}
\description{
Implement grid search for the bayes object
}
