% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant.R
\name{grp_indices}
\alias{grp_indices}
\title{Implementations related to Discriminant analysis
Create indices for each class K}
\usage{
grp_indices(groups, Y)
}
\arguments{
\item{groups}{factor levels for all K classes}

\item{Y}{target vector}
}
\value{
list of group indices within target vector
}
\description{
Implementations related to Discriminant analysis
Create indices for each class K
}
