% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{huber_loss}
\alias{huber_loss}
\title{Huber Loss function (smooth mean absolute error) with parameter delta}
\usage{
huber_loss(e, delta)
}
\arguments{
\item{e}{vector of error terms Yi - Y^hat_i}

\item{delta}{const for determining huber loss conditional}
}
\value{
huber loss function error
}
\description{
Huber Loss function (smooth mean absolute error) with parameter delta
}
\examples{
e <- matrix(data=0, nrow=10,ncol=1)
delta <- 1.0
huber_loss(e, delta)

e <- rnorm(10, 0, 1)
huber_loss(e, delta)

}
