% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kblock_kern}
\alias{kblock_kern}
\title{K-Block Kernel}
\usage{
kblock_kern(i, j, m, k)
}
\arguments{
\item{i}{row position in matrix}

\item{j}{column position in matrix}

\item{m}{boolean spatial matrix}

\item{k}{integer, size of window}
}
\value{
TRUE if window exists, otherwise FALSE
}
\description{
Special case of a kernel used for estimating spatial
relationships to assess similarity of geographically
proximate neighbors.
}
\details{
Current implementation is used in Bayesian classification.
NOTE that the author is currently unaware of this
kernel being available in the literature.

For more details see the help vignette:
\code{vignette("homophily-kernel", package = "supml")}
}
