% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{kernelDensity.bayes}
\alias{kernelDensity.bayes}
\title{Gaussian Kernel Density Estimate for Naive Bayes}
\usage{
\method{kernelDensity}{bayes}(obj, x, mu = 0)
}
\arguments{
\item{obj}{requires instantiated Bayes object}

\item{x}{continuous feature vector}
}
\value{
fhat, point estimate with gaussian kernel
}
\description{
Implements HT (6.23)
}
