% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant.R
\name{ldf}
\alias{ldf}
\title{Linear discriminant function (LDF) implementation}
\usage{
ldf(X, y, kpi, kmu, sig)
}
\description{
LDF is computed for k in K classes for LDA. See
HT, Elements of Statistical Learning equation (4.10)
}
