% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{least_squares_huber_gradient}
\alias{least_squares_huber_gradient}
\title{Least Squares with Huber Loss Function gradient}
\usage{
least_squares_huber_gradient(X, y, theta, alpha, ...)
}
\description{
Least Squares with Huber Loss Function gradient
}
