% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{least_squares_l1_gradient}
\alias{least_squares_l1_gradient}
\title{Least Squares with Mean Absolute Error (L1 norm) gradient}
\usage{
least_squares_l1_gradient(X, y, theta, alpha)
}
\description{
Least Squares with Mean Absolute Error (L1 norm) gradient
}
