% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{least_squares_ql_gradient}
\alias{least_squares_ql_gradient}
\title{Least Squares with Quadratic Loss gradient}
\usage{
least_squares_ql_gradient(X, y, theta, alpha)
}
\description{
Related to Mean Squared Error, L2 norm
}
