% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{lgm_gradient}
\alias{lgm_gradient}
\title{Logistic gradient computation}
\usage{
lgm_gradient(w, X, t)
}
\arguments{
\item{w}{vector of weights}

\item{X}{matrix of predictors}

\item{t}{vector of target}
}
\value{
w vector of weights
}
\description{
Logistic gradient computation
}
