% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{lgm_yhat}
\alias{lgm_yhat}
\title{Implementations for Logistic Regression
Compute yhat for logistic regression}
\usage{
lgm_yhat(w, X)
}
\arguments{
\item{w}{vector of weights}

\item{X}{matrix of predictors}
}
\value{
dot product of X and w with sigmoid function
}
\description{
Implementations for Logistic Regression
Compute yhat for logistic regression
}
