% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant.R
\name{linda}
\alias{linda}
\title{Linear Discriminant Analysis implementation}
\usage{
linda(X, y)
}
\arguments{
\item{X}{matrix of predictors (no intercept)}

\item{y}{target vector}
}
\value{
yhat vector
}
\description{
Linear Discriminant Analysis implementation
}
