% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{local_lm}
\alias{local_lm}
\title{Local Linear Regression}
\usage{
local_lm(X, y, lambda = 1.5)
}
\arguments{
\item{X}{attribute vector}

\item{y}{target vector}
}
\value{
y hat target vector
}
\description{
Referencing HT (6.7)
}
