% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic_gradient}
\alias{logistic_gradient}
\title{Compute the logistic gradient with learning rate alpha}
\usage{
logistic_gradient(X, t, w, alpha)
}
\arguments{
\item{X}{matrix of predictors}

\item{t}{vector of target}

\item{w}{vector of weights}

\item{alpha}{float, learning rate}
}
\value{
w vector of weights
}
\description{
Compute the logistic gradient with learning rate alpha
}
\examples{

}
