% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\docType{data}
\name{LOGMIN}
\alias{LOGMIN}
\title{Implement Naive Bayes Classifier}
\format{An object of class \code{numeric} of length 1.}
\usage{
LOGMIN
}
\description{
Implement Naive Bayes Classifier
}
\keyword{datasets}
