% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{mae}
\alias{mae}
\title{Compute Mean Absolute Error (L1 norm)}
\usage{
mae(e, delta = 0)
}
\arguments{
\item{e}{vector of error terms Yi - Y^hat_i}
}
\value{
mean absolute error
}
\description{
Compute Mean Absolute Error (L1 norm)
}
\examples{
e <- matrix(data=0, nrow=10,ncol=1)
mae(e)

e <- rnorm(10, 0, 1)
mae(e)

}
