% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{matrix_group}
\alias{matrix_group}
\title{Convert matrix to visualization format}
\usage{
matrix_group(X)
}
\arguments{
\item{X}{boolean square matrix}
}
\value{
dataframe containing group
information with factors
}
\description{
This function is meant to be used when
creating an overlaying layer when
generating maps in ggplot2.
}
