% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Testing data documentation
mtcars dataset}
\format{An object of class \code{data.frame} with 32 rows and 11 columns.}
\usage{
data(mtcars)
}
\description{
Data from the mtcars dataset included in R
by default.
}
\keyword{datasets}
