% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric object}
\usage{
normalize(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
normalized values (0,1)
}
\description{
Normalize a numeric object
}
