% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{plotgrid}
\alias{plotgrid}
\title{Coordinate grid for plotting}
\usage{
plotgrid(m, n, step)
}
\arguments{
\item{m}{number of rows}

\item{n}{number of columns}

\item{step}{step size, ex. sqrt(m) / (1.0 / n)}
}
\description{
Matrix must be square
}
