% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant.R
\name{pooled_covar}
\alias{pooled_covar}
\title{Compute Pooled Covariance Matrix}
\usage{
pooled_covar(X, y)
}
\arguments{
\item{X}{matrix of predictors}

\item{y}{target vector}
}
\value{
pooled p by p covariance matrix
}
\description{
Equation comes from Pooled Variance
https://en.wikipedia.org/wiki/Pooled_variance
}
