% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{popmat}
\alias{popmat}
\title{Populate a Spatial Matrix with Aggregated Values}
\usage{
popmat(xf, yf, z, m)
}
\arguments{
\item{xf}{factor vector of x coordinate}

\item{yf}{factor vector of y coordinate}

\item{z}{numeric vector of aggregated measure}

\item{m}{matrix with levels(xf) by levels(yf) dimensions}
}
\value{
m spatial matrix
}
\description{
Once a square spatial matrix has been created using
normalized coordinates, we must populate it with
values of interest.
}
