% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{postSpatialProbs.bayes}
\alias{postSpatialProbs.bayes}
\title{Posterior Spatial probabilties}
\usage{
\method{postSpatialProbs}{bayes}(object, X)
}
\arguments{
\item{object}{s3 bayes object}

\item{X}{spatial feature vector}
}
\value{
matrix of spatial features
}
\description{
Multinomial classification is not currently supported.
Assume m\link{i,j} is TRUE if FALSE, then
check to see if K-Block kernel evaluates to TRUE.
If K-Block is TRUE, then use conditional probabilities
to determine class membership.
}
