% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{predict.bayes}
\alias{predict.bayes}
\title{Predict log probabilities for each class K}
\usage{
\method{predict}{bayes}(object, X)
}
\arguments{
\item{object}{s3 object bayes}

\item{X}{feature matrix}
}
\value{
s3 object bayes with prediction attribute
}
\description{
Predict log probabilities for each class K
}
