% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{predict_proba.bayes}
\alias{predict_proba.bayes}
\title{Predicts the log probabilities for each class and feature}
\usage{
\method{predict_proba}{bayes}(obj, X)
}
\arguments{
\item{obj}{s3 object, bayes}

\item{X}{feature vector}
}
\value{
bayes s3 object
}
\description{
Predicts the log probabilities for each class and feature
}
