% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{priorProbs.bayes}
\alias{priorProbs.bayes}
\title{Prior probability fitting step for Continous/Categorical data}
\usage{
\method{priorProbs}{bayes}(object, y)
}
\arguments{
\item{object}{bayesian s3 object}

\item{y}{target vector, vector}

\item{natural}{log probabilities for each class y given |y|}
}
\description{
Compute prior probs for each class in y. This method
should only be used by customizing your own Bayesian
classifier.
}
