% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{score.bayes}
\alias{score.bayes}
\title{Accuracy scoring function for Binomial Naive Bayes}
\usage{
\method{score}{bayes}(obj, y)
}
\arguments{
\item{obj}{bayes s3 object}

\item{y}{target vector for training set}
}
\value{
float, accuracy score
}
\description{
Converts log probabilities to probabilities before
scoring.
}
