% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradesc.R
\name{sgd}
\alias{sgd}
\title{Stochastic Gradient Descent}
\usage{
sgd(alpha, X, y, grad, N, precision = 1e-04)
}
\arguments{
\item{alpha}{learning rate constant}

\item{X}{matrix of predictors including 1 for beta 0}

\item{y}{vector of dependent variable}

\item{grad}{gradient function}

\item{N}{maximum number of iterations}

\item{precision}{constant used for approximation condition}
}
\value{
list object with theta, convergence status, and number of iterations
}
\description{
Stochastic Gradient Descent
}
\examples{
x <- runif(1000, -5, 5)
y <- x + rnorm(1000) + 3
X <- cbind(1, matrix(x))
h <- sgd(0.1, X, y, least_squares_gradient, 1000)
h$theta
h$converged == TRUE
h$iterations

}
