% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{spatial_fit.bayes}
\alias{spatial_fit.bayes}
\title{Fit spatial data}
\usage{
\method{spatial_fit}{bayes}(object, X, Y)
}
\arguments{
\item{object}{bayes s3 object}

\item{X}{boolean spatial feature matrix}

\item{Y}{boolean matrix of same dimensions as X}
}
\description{
The conditional priors with a maximum likelihood
are fit to the model using spatial data.
}
