% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{spatialProbs.bayes}
\alias{spatialProbs.bayes}
\title{Conditional probability fitting step for Spatial data}
\usage{
\method{spatialProbs}{bayes}(object, X, y)
}
\arguments{
\item{object}{bayesian s3 object with mappings set}

\item{X}{boolean spatial feature matrix}

\item{y}{target vector, factor}
}
\value{
natural log probabilities for each class in y
}
\description{
This method should only be used for customizing your
own Bayesian classifier. Computes p(x_ij|k), p(x_ij|w)
where k = TRUE in k-block.
}
