% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{w}
\alias{w}
\title{mtcars derivative variables}
\format{An object of class \code{numeric} of length 32.}
\usage{
data(x,w,y)
}
\description{
Data derived from mtcars for testing
Gaussian kernel. Tests fail if I don't
overexplain how data is generated for
tests.
}
\details{
Generated:

y = as.matrix(mtcars$mpg)
x = as.matrix(mtcars$wt)
w = runif(length(x))
}
\keyword{datasets}
