% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{wls}
\alias{wls}
\title{Implement local linear regression with a Gaussian kernel,
as a function of tuning parameter lambda.
Weighted Least Squares}
\usage{
wls(x, y, w)
}
\arguments{
\item{y}{target vector}

\item{w}{weight vector}

\item{X}{feature matrix without intercept column}
}
\value{
coefficients theta
}
\description{
Analytic solution using Strang 5th ed.
}
