% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes}
\alias{bayes}
\title{Bayesian object}
\usage{
bayes(map)
}
\arguments{
\item{map}{list of variable types for X}
}
\value{
s3 object, bayes
}
\description{
Used for Bayesian classification with kernel
density estimation
}
