% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{kernelDensity.bayes}
\alias{kernelDensity.bayes}
\title{Kernel Density Estimate for Naive Bayes}
\usage{
\method{kernelDensity}{bayes}(object, x, kernel, ...)
}
\arguments{
\item{object}{requires instantiated Bayes object}

\item{x}{feature vector}

\item{kernel}{name of kernel ("gaussian", "kblock")}

\item{...}{kernel specific arguments}
}
\value{
kernel estimate
}
\description{
Implements HT (6.23), implements K-Block kernel
}
